+++ tools/cookbook/main.go
@@ -155,7 +155,7 @@ func markerAbove(lines []string, i int) bool {
 var countWords = map[string]int{
 	"One": 1, "Two": 2, "Three": 3, "Four": 4, "Five": 5, "Six": 6,
 	"Seven": 7, "Eight": 8, "Nine": 9, "Ten": 10, "Eleven": 11, "Twelve": 12,
-	"Thirteen": 13, "Fourteen": 14, "Fifteen": 15, "Sixteen": 16,
+	"Thirteen": 13, "Fourteen": 14, "Fifteen": 15, "Sixteen": 16, "Seventeen": 17,
 }
 
 var countLine = regexp.MustCompile(`(?m)^([A-Z][a-z]+) recipes,`)